requirements.txt
flask
python-dotenv
sqlalchemy
psycopg2-binary
langchain
langchain-groq